added to todomaster
Need to disable manual edit and adds on all datagrids (eliminates the useless empty line at the bottom of the datagrid view. This will also allow for select all to work properly in datagrids. to set this property set allow user add to false, as well as allow user delete(or remove) and allow user edit all to false in dgv properties.